srv/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
celery==5.4.0
redis==5.0.4
boto3==1.34.131
python-multipart==0.0.9
